move stories to a dedicated folder
